task 004 product

+++ Task004_product/Program.cs
@@ -1,7 +1,7 @@
 ﻿//Написать программу вычисления произведения чисел от 1 до N
 Console.WriteLine("Input positive integer and press Enter: ");
 int n = int.Parse(Console.ReadLine());
-int product = 1;
+double product = 1;
 for (int count = 1; count<=n; count++)
 {
     product = product * count;
